+++ example/example_methods.go
@@ -132,4 +132,3 @@ func RepeatString(s string, n int32) plgo.SetOf[string] {
 	}
 	return result
 }
-